requirements.txt
flask
pillow
numpy
werkzeug
